[2.46] don't schedule activity state changes
This change is needed for Page Lifecycle API. Without this change
when we call blur->hide or show->focus in a row then we will get only
one setActivityState message with both changes. It causes that
the browser dispatches blur/focus events before visibilityChange,
which is not desired in case of show->focus. Additionally calling
blur->hide->freeze on UI side in a row dispatches freeze event before
blur and visibilityChange events on WebProcess, because freeze is
called directly, while setActivityState is scheduled.